requirements.txt
PySocks==1.7.1
python-dotenv==0.10.3
python-telegram-bot==12.1.1
